+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - quarantining is now considered absolutely; public activities are no longer
   an exception.
+- also merged quarantine and mrf reject - quarantine is now deprecated
 - flavours:
   - amd64 is built for debian stable. Compatible with ubuntu 20.
   - ubuntu-jammy is built for... well, ubuntu 22 (LTS)